Multi-session with same Minecraft account with specific permission
Players with permission bungeecord.multiple_connect can have multiples
connections with the same Minecraft account. To manage UUID and Name conflict,
an increment value is appended to the player name, and the uuid used is the
offline one. So be carefull: Spigot servers and plugins will considers 2
players from the same account like if they were 2 different MC account.